backend/requirements.txt
```
fastapi==0.115.0
uvicorn==0.30.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
spacy==3.7.6
boto3==1.34.162
```